Start modular split
