+++ buffer.h
@@ -1,0 +1,15 @@
+#ifndef BUFFER_H
+#define BUFFER_H
+
+#include "text.h"
+
+struct files {
+    char *name;
+    struct files *next_name;
+};
+
+struct text *txtalloc(void);
+struct files *name_alloc(void);
+ee_char *resiz_line(int factor, struct text *rline, int rpos);
+
+#endif /* BUFFER_H */
